# Reduce calls to RandomAccessFile.setLength in NIOFileHandle