Instead of calling ```raf.setLength``` with the exact length required for the current write, this increases the length by ```NIOFileHandle.getBufferSize()``` in anticipation of future writes.  ```NIOFileHandle.close()``` then reconciles any discrepancy between the actual and expected file lengths, so that there are no extra bytes in the file.  This was the least invasive change I could think of, but happy to discuss if anyone has other thoughts.

See https://trac.openmicroscopy.org/ome/ticket/6931 and https://trello.com/c/OimiHAQY/175-profile-tiff-writing-on-windows.  This does have a noticeable impact on the performance of ```bfconvert``` on Windows; in the test case on the ticket, local Windows times are now down to ~52s vs ~95s.