@sbesson 2017-09-19
Sorry for the delay @melissalinkert, slowly getting back to this. Is there a minimal example consuming the `loci.common.NIOFileHandle` API that could be run on Linux and Windows with and without this PR to test its effect in isolation of Bio-Formats?