@melissalinkert 2017-09-19
I'd expect that something as simple as:

```
NIOFileHandle out = new NIOFileHandle("test", "rw");
long byteCount = 1024 * 1024; // 1 MB
for (long i=0; i<byteCount; i++) {
  out.write(1);
}
out.close();
```

would be noticeably faster with this PR.