@dgault 2017-09-21
The code looks good and the solution is effective without being a major overhaul.

Testing using the sample code suggested provides a substantial improvement in performance. 

Without the PR:
```
1024 * 1024: Time elapsed: 11898 ms
2048 * 2048: Time elapsed: 36811 ms
4096 * 4096: Time elapsed: 147633 ms
```

With the PR:
```
1024 * 1024: Time elapsed: 2089 ms
2048 * 2048: Time elapsed: 7692 ms
4096 * 4096: Time elapsed: 31783 ms
```