@sbesson 2017-09-25
Using the following `Makefile` to build Bio-Formats with and without this PR:

```
BUILDDIR=/tmp/build

clean:
	rm -rf $(BUILDDIR)
	mkdir -p $(BUILDDIR)
	test -f bioformats-5.7.0.zip || wget https://downloads.openmicroscopy.org/bio-formats/5.7.0/artifacts/bioformats-5.7.0.zip
	test -f setlength-performance.zip || wget https://github.com/melissalinkert/ome-common-java/archive/setlength-performance.zip

bf1:
	unzip bioformats-5.7.0.zip && mv bioformats-5.7.0 $(BUILDDIR)/bf1
	ant tools -f $(BUILDDIR)/bf1/build.xml -Dmaven.repo.local=$(BUILDDIR)/.m2/repository

ome-common:
	unzip setlength-performance.zip && mv ome-common-java-setlength-performance $(BUILDDIR)/ome-common
	mvn clean install -f $(BUILDDIR)/ome-common/pom.xml -Dmaven.repo.local=$(BUILDDIR)/.m2/repository

bf2:
	unzip bioformats-5.7.0.zip && mv bioformats-5.7.0 $(BUILDDIR)/bf2
	sed -i '' -e 's/5.3.2/5.3.3-SNAPSHOT/g' $(BUILDDIR)/bf2/pom.xml
	ant tools -f $(BUILDDIR)/bf2/build.xml -Dmaven.repo.local=$(BUILDDIR)/.m2/repository -Dome-common.version=5.3.3-SNAPSHOT
     
.PHONY: clean bf1 ome-common bf2
```

and running `/tmp/build/bf{1,2}/tools/showinf -nopix -cache` on a few filesets, I did not see any invalidated cache files.

From an integration perspective, this makes this PR eligible for an inclusion in a patch release of Bio-Formats. The main issue is the ongoing discussion on #12 (see https://trello.com/c/Ym1HWlQG/159-ome-common-location-exception) and the fact more discussion/configuration might be required before propagating the latter changes.

In order to move forward, my proposal would be to revert #12, merge this into `master` and cut a patch release `ome-common:5.3.3` with this performance improvement. We can then propagate the release down the stack up to Bio-Formats. Would that work for you @melissalinkert @dgault ?